fix: update folder structure to match defined rules
