+++ src/components/atoms/index.tsx
@@ -1,7 +1,5 @@
 /* eslint-disable react-refresh/only-export-components */
 export * from "./button";
-export { IconButton } from "./button/icon-button";
-export type { IconButtonProps } from "./button/icon-button";
 export * from "./info-box";
 export * from "./label-chip";
 export { Toast, useToast, Snackbar, useSnackbar } from "./snackbar";
