+++ src/components/atoms/radio-button/types.ts
@@ -1,0 +1,10 @@
+export interface RadioButtonProps {
+  label?: string;
+  value: string;
+  checked?: boolean;
+  onChange?: (value: string) => void;
+  disabled?: boolean;
+  helperText?: string;
+  labelClassName?: string;
+  helperTextClassName?: string;
+}
