+++ src/components/molecules/score-increase-decrease/types.ts
@@ -1,0 +1,22 @@
+export type ScoreIncreaseDecreaseSize = "small" | "medium" | "large";
+
+export const SCORE_INCREASE_DECREASE_SIZES: ScoreIncreaseDecreaseSize[] = [
+  "small",
+  "medium",
+  "large",
+];
+
+export interface ScoreIncreaseDecreaseProps {
+  score: number;
+  onIncrease: () => void;
+  onDecrease: () => void;
+  onScoreChange?: (score: number) => void;
+  variant?: ScoreIncreaseDecreaseSize;
+  disabled?: boolean;
+  danger?: boolean;
+  inFocus?: boolean;
+  editable?: boolean;
+  scoreInputEditable?: boolean;
+  min?: number;
+  max?: number;
+}
